+++ dao/src/main/java/com/eghm/dao/mapper/SysOperatorRoleMapper.java
@@ -22,7 +22,7 @@ public interface SysOperatorRoleMapper {
      *
      * @param id
      */
-    SysOperatorRole selectByPrimaryKey(Integer id);
+    SysOperatorRole selectByPrimaryKey(Long id);
 
     /**
      * 根据主键来更新部分数据库记录
@@ -37,14 +37,14 @@ public interface SysOperatorRoleMapper {
      * @param operatorId 角色id
      * @return 角色id列表
      */
-    List<Integer> getByOperatorId(@Param("operatorId") Integer operatorId);
+    List<Integer> getByOperatorId(@Param("operatorId") Long operatorId);
 
     /**
      * 删除用户所有的角色
      * @param operatorId 管理人员id
      * @return 影响条数
      */
-    int deleteByOperatorId(@Param("operatorId") Integer operatorId);
+    int deleteByOperatorId(@Param("operatorId") Long operatorId);
 
     /**
      * 批量添加用户角色关系信息
@@ -52,5 +52,5 @@ public interface SysOperatorRoleMapper {
      * @param roleList 角色列表
      * @return 条数
      */
-    int batchInsertOperatorRole(@Param("operatorId") Integer operatorId, @Param("roleList") List<Integer> roleList);
+    int batchInsertOperatorRole(@Param("operatorId") Long operatorId, @Param("roleList") List<Long> roleList);
 }
