+++ dao/src/main/java/com/eghm/dao/mapper/SysRoleMapper.java
@@ -23,7 +23,7 @@ public interface SysRoleMapper {
      *
      * @param id
      */
-    SysRole selectByPrimaryKey(Integer id);
+    SysRole selectByPrimaryKey(Long id);
 
     /**
      * 根据主键来更新部分数据库记录
@@ -44,21 +44,21 @@ public interface SysRoleMapper {
      * @param operatorId 用户id
      * @return 角色列表
      */
-    List<SysRole> getRoleList(@Param("operatorId") Integer operatorId);
+    List<SysRole> getRoleList(@Param("operatorId") Long operatorId);
 
     /**
      * 获取角色拥有的菜单列表
      * @param roleId 角色id
      * @return 菜单menuIds
      */
-    List<Integer> getRoleMenu(@Param("roleId") Integer roleId);
+    List<Integer> getRoleMenu(@Param("roleId") Long roleId);
 
     /**
      * 删除角色菜单关联信息 物理删除
      * @param roleId 角色id
      * @return 删除条件
      */
-    int deleteRoleMenu(@Param("roleId") Integer roleId);
+    int deleteRoleMenu(@Param("roleId") Long roleId);
 
     /**
      * 批量添加角色菜单关联信息
@@ -66,5 +66,5 @@ public interface SysRoleMapper {
      * @param menuIdList 菜单列表
      * @return 查询条数
      */
-    int batchInsertRoleMenu(@Param("roleId") Integer roleId, @Param("menuIdList") List<Integer> menuIdList);
+    int batchInsertRoleMenu(@Param("roleId") Long roleId, @Param("menuIdList") List<Long> menuIdList);
 }
