+++ model/src/main/java/com/eghm/model/dto/ext/ActionRecord.java
@@ -27,6 +27,6 @@ public class ActionRecord implements Serializable {
      * 操作人id
      */
     @BackstageTag
-    private Integer operatorId;
+    private Long operatorId;
 
 }
