+++ model/src/main/java/com/eghm/model/dto/ext/ApiHolder.java
@@ -41,7 +41,7 @@ public class ApiHolder {
     /**
      * 强制获取用户信息
      */
-    public static Integer getUserId() {
+    public static Long getUserId() {
         return getUserId(true);
     }
 
@@ -49,7 +49,7 @@ public class ApiHolder {
      * 尝试获取用户信息
      * @return 如果用户未登陆获取结果为0
      */
-    public static Integer tryGetUserId() {
+    public static Long tryGetUserId() {
         return getUserId(false);
     }
 
@@ -58,8 +58,8 @@ public class ApiHolder {
      * @param need 用户是否需要登陆,如果必须登陆登陆,用户id不存在则抛异常
      * @return 用户id
      */
-    private static Integer getUserId(boolean need) {
-        Integer userId = get().getUserId();
+    private static Long getUserId(boolean need) {
+        Long userId = get().getUserId();
         if (need && userId == null) {
             throw new BusinessException(ErrorCode.USER_LOGIN_TIMEOUT);
         }
